requirements.txt
google-api-python-client
sqlalchemy
psycopg2
fastapi
django-background-tasks
django